stack/packages/functions/requirements.txt
aws-lambda-powertools[validation]
jsonpickle
simplejson
python-jose[cryptography]
aws_requests_auth
